Update Specta
Summary:
This was orignally pinned to a commit directly to fix builds on Xcode 7 as part of D8570. This is no longer necessary.

By changing this back to a tag, it will allow us to work around an issue in Carthage 0.12.0 as well, see https://github.com/Carthage/Carthage/issues/1080

Version: patch

Test Plan: N/A

Reviewers: sylvain, robb, engin, westin, James

Reviewed By: James

Subscribers: jenkins

Differential Revision: https://phabricator.automatic.co/D11480